address review: move defhands parser module + list-style options
