+++ docs/config.adoc
@@ -2289,22 +2289,33 @@ a key cannot appear in both groups; partial definitions (only one hand) are allo
 unlisted keys have no hand assignment.
 
 `tap-hold-opposite-hand` takes `$timeout $tap-action $hold-action` followed by
-optional keyword pairs. It requires `defhands` to be defined.
+optional option lists in the form `(name value)`. It requires `defhands` to be defined.
 Internally it uses the same `HoldTapConfig::Custom` machinery as `tap-hold-release-keys`.
 
 [cols="1,1,1,2"]
 |===
 | Option | Values | Default | Description
 
-| `:timeout` | `tap`, `hold` | `tap` | Action when hold timeout elapses (optimized for home-row-mod behavior).
-| `:same-hand` | `tap`, `hold`, `ignore` | `tap` | Action when a same-hand key is pressed.
-| `:neutral-keys` | `(key list)` | (empty) | Keys treated as neutral, overriding their `defhands` assignment.
-| `:neutral` | `tap`, `hold`, `ignore` | `ignore` | Action when a `:neutral-keys` key is pressed.
-| `:unknown-hand` | `tap`, `hold`, `ignore` | `ignore` | Action when a key not in `defhands` is pressed.
+| `(timeout <value>)` | `tap`, `hold` | `tap` | Action when hold timeout elapses (optimized for home-row-mod behavior).
+| `(same-hand <value>)` | `tap`, `hold`, `ignore` | `tap` | Action when a same-hand key is pressed.
+| `(neutral-keys (key list))` | `(key list)` | (empty) | Keys treated as neutral, overriding their `defhands` assignment.
+| `(neutral <value>)` | `tap`, `hold`, `ignore` | `ignore` | Action when a `neutral-keys` key is pressed.
+| `(unknown-hand <value>)` | `tap`, `hold`, `ignore` | `ignore` | Action when a key not in `defhands` is pressed.
 |===
 
 When a value is `ignore`, that key press is skipped and the action
 waits for the next key or timeout.
+
+Example with options:
+[source]
+----
+(defalias
+  fctl (tap-hold-opposite-hand 180 f lctl
+    (same-hand ignore)
+    (timeout hold)
+    (neutral-keys (spc tab))
+    (neutral tap)))
+----
 
 See `cfg_samples/opposite-hand-hrm.kbd` for a full working example.
